requirements.txt
flask==2.2.3
python-dotenv==1.0.0
peewee==3.16.0
psycopg2-binary==2.9.6
redis==4.5.4
rq==1.13.0
dacite==1.8.0